add header for AsyncDAGNet
Summary: This diff adds a header file for net_gpu.cc so that the AsyncDAGNet class can be used to create other derived classes.

Reviewed By: ajtulloch

Differential Revision: D4230046

fbshipit-source-id: 379c3ff7ebb7aeeb4294f39e6f5d1ecad48b92f0
